ShopDemo/CustomHelpers/EncodeStringHelpers.cs
using System;
using System.Text;

namespace ShopDemo.CustomHelpers
{
    public static class EncodeStringHelpers
    {
        public static string ConvertStringToUtf8(string value)
        {
            return "=?UTF-8?B?" + Convert.ToBase64String(Encoding.UTF8.GetBytes(value)) + "?=";
        }

        public static string ToBase64(this string value)
        {
            byte[] bytes = Encoding.UTF8.GetBytes(value);
            return Convert.ToBase64String(bytes);
        }

        public static string FromBase64(this string value)
        {
            byte[] bytes = Convert.FromBase64String(value);
            return Encoding.UTF8.GetString(bytes);
        }
    }
}